Patch LayerNorm and normalise
There is a problem in normalise that if `std(x) \approx 0`, then the
chain rule evaluates to NaN. See e.g. here: [https://github.com/FluxML/Flux.jl/issues/2096].

We tried to fix this here by adding some noise to x, although that might
not be the best solution. We also fix in a later commit that all images
actually have some noise in the background.